support -1 as the group with all children

+++ graalpython/com.oracle.graal.python/src/com/oracle/graal/python/runtime/PosixResources.java
@@ -91,11 +91,16 @@ public class PosixResources {
         @Override
         public int waitFor() throws InterruptedException {
             for (Process child : children) {
-                if (child != null) {
-                    return child.waitFor();
+                if (child != null && child != this) {
+                    int exitCode = child.waitFor();
+                    int childIndex = children.indexOf(child);
+                    if (childIndex > 0) {
+                        children.set(childIndex, null);
+                    }
+                    return exitCode;
                 }
             }
-            return 0;
+            throw new IndexOutOfBoundsException();
         }
 
         @Override
@@ -116,7 +121,7 @@ public class PosixResources {
         @Override
         public int exitValue() {
             for (Process child : children) {
-                if (!child.isAlive()) {
+                if (child != this && !child.isAlive()) {
                     return child.exitValue();
                 }
             }
@@ -126,7 +131,7 @@ public class PosixResources {
         @Override
         public void destroy() {
             for (Process child : children) {
-                if (child != null) {
+                if (child != null && child != this) {
                     child.destroy();
                 }
             }
@@ -135,7 +140,7 @@ public class PosixResources {
         @Override
         public Process destroyForcibly() {
             for (Process child : children) {
-                if (child != null) {
+                if (child != null && child != this) {
                     child.destroyForcibly();
                 }
             }
@@ -395,8 +400,12 @@ public class PosixResources {
     }
 
     private Process getChild(int pid) throws IndexOutOfBoundsException {
-        if (pid < 0) {
+        if (pid < -1) {
             throw new IndexOutOfBoundsException("we do not support process groups");
+        } else if (pid == -1) {
+            // -1 - any child process.
+            // 0 - any child process with the same process group.
+            return children.get(0);
         } else {
             return children.get(pid);
         }
@@ -423,7 +432,9 @@ public class PosixResources {
     public int waitpid(int pid) throws IndexOutOfBoundsException, InterruptedException {
         Process process = getChild(pid);
         int exitStatus = process.waitFor();
-        children.set(pid, null);
+        if (pid > 0) { // cannot delete process groups
+            children.set(pid, null);
+        }
         return exitStatus;
     }
